# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ChatNet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     * @pattern ^[a-zA-Z0-9_]+$
     */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  /** 8-digit friend token (xx.xx.xx.xx format) */
  id: string;
  password: string;
}

export interface GuestLoginInput {
  /**
     * @minLength 1
     * @maxLength 30
     */
  displayName: string;
}

export interface UpdateMeInput {
  /**
     * @minLength 1
     * @maxLength 30
     */
  displayName?: string;
  quietMode?: boolean;
}

export interface AuthUser {
  id: string;
  /** What other users see — username for registered, chosen name for guests */
  displayName: string;
  /**
     * 8-digit token in xx.xx.xx.xx format. Null for guests.
     * @nullable
     */
  friendToken?: string | null;
  isGuest: boolean;
  quietMode: boolean;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface GeneralMessage {
  id: string;
  content: string;
  senderId: string;
  /** Display name of the sender */
  senderLabel: string;
  /**
     * Friend token of sender, null for guests
     * @nullable
     */
  senderToken?: string | null;
  createdAt: string;
  seenByMe: boolean;
}

export interface DirectMessage {
  id: string;
  content: string;
  senderId: string;
  fromMe: boolean;
  createdAt: string;
  /** @nullable */
  seenAt?: string | null;
}

export interface Friend {
  id: string;
  /** Display name of the friend */
  label: string;
  /**
     * Friend token, null if the friend is a guest
     * @nullable
     */
  friendToken?: string | null;
  createdAt: string;
  unreadCount?: number;
}

export interface AddFriendInput {
  /** 8-digit token in xx.xx.xx.xx format */
  token: string;
}

export interface SeenResult {
  markedCount: number;
}

export interface RemoveFriendResult {
  removed: boolean;
}

export type GetGeneralMessagesParams = {
limit?: number;
/**
 * Message ID to paginate before
 * @nullable
 */
before?: string | null;
};