# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChatNet API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: general
    description: General chat (whole server)
  - name: friends
    description: Friend management and DM chat
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username already taken

  /auth/guest:
    post:
      operationId: guestLogin
      tags: [auth]
      summary: Enter as an anonymous guest with a chosen display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestLoginInput"
      responses:
        "201":
          description: Guest session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid display name

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update display name and/or quiet mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
        "409":
          description: Username already taken

  /general/messages:
    get:
      operationId: getGeneralMessages
      tags: [general]
      summary: Get general chat messages (paginated)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: ["string", "null"]
          description: Message ID to paginate before
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneralMessage"

  /friends:
    get:
      operationId: getFriends
      tags: [friends]
      summary: Get friend list
      responses:
        "200":
          description: Friend list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friend"

  /friends/add:
    post:
      operationId: addFriend
      tags: [friends]
      summary: Add a friend by 8-digit token (xx.xx.xx.xx)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFriendInput"
      responses:
        "201":
          description: Friend added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
        "400":
          description: Invalid token or already friends
        "403":
          description: User has quiet mode enabled
        "404":
          description: Token not found

  /friends/{friendId}/messages:
    get:
      operationId: getFriendMessages
      tags: [friends]
      summary: Get DM messages with a friend (last 50)
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DM messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DirectMessage"
        "403":
          description: Not friends

  /friends/{friendId}/seen:
    post:
      operationId: markSeen
      tags: [friends]
      summary: Mark all DMs from a friend as seen
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked seen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeenResult"

  /friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Friend removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveFriendResult"
        "404":
          description: Friendship not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_]+$"
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [id, password]
      properties:
        id:
          type: string
          description: "8-digit friend token (xx.xx.xx.xx format)"
        password:
          type: string

    GuestLoginInput:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 30

    UpdateMeInput:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 30
        quietMode:
          type: boolean

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"

    AuthUser:
      type: object
      required: [id, displayName, isGuest, quietMode]
      properties:
        id:
          type: string
        displayName:
          type: string
          description: "What other users see — username for registered, chosen name for guests"
        friendToken:
          type: ["string", "null"]
          description: "8-digit token in xx.xx.xx.xx format. Null for guests."
        isGuest:
          type: boolean
        quietMode:
          type: boolean

    GeneralMessage:
      type: object
      required: [id, content, senderId, senderLabel, createdAt, seenByMe]
      properties:
        id:
          type: string
        content:
          type: string
        senderId:
          type: string
        senderLabel:
          type: string
          description: "Display name of the sender"
        senderToken:
          type: ["string", "null"]
          description: "Friend token of sender, null for guests"
        createdAt:
          type: string
          format: date-time
        seenByMe:
          type: boolean

    DirectMessage:
      type: object
      required: [id, content, senderId, fromMe, createdAt]
      properties:
        id:
          type: string
        content:
          type: string
        senderId:
          type: string
        fromMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
        seenAt:
          type: ["string", "null"]
          format: date-time

    Friend:
      type: object
      required: [id, label, createdAt]
      properties:
        id:
          type: string
        label:
          type: string
          description: "Display name of the friend"
        friendToken:
          type: ["string", "null"]
          description: "Friend token, null if the friend is a guest"
        createdAt:
          type: string
          format: date-time
        unreadCount:
          type: integer

    AddFriendInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: "8-digit token in xx.xx.xx.xx format"

    SeenResult:
      type: object
      required: [markedCount]
      properties:
        markedCount:
          type: integer

    RemoveFriendResult:
      type: object
      required: [removed]
      properties:
        removed:
          type: boolean