# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ChatNet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  /** What other users see — username for registered, chosen name for guests */
  displayName: string;
  /**
     * 8-digit token in xx.xx.xx.xx format. Null for guests.
     * @nullable
     */
  friendToken?: string | null;
  isGuest: boolean;
  quietMode: boolean;
}