# lib/api-zod/src/generated/types/friend.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ChatNet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Friend {
  id: string;
  /** Display name of the friend */
  label: string;
  /**
     * Friend token, null if the friend is a guest
     * @nullable
     */
  friendToken?: string | null;
  createdAt: Date;
  unreadCount?: number;
}