# lib/api-zod/src/generated/types/generalMessage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ChatNet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneralMessage {
  id: string;
  content: string;
  senderId: string;
  /** Display name of the sender */
  senderLabel: string;
  /**
     * Friend token of sender, null for guests
     * @nullable
     */
  senderToken?: string | null;
  createdAt: Date;
  seenByMe: boolean;
}