# lib/api-zod/src/generated/types/updateMeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ChatNet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMeInput {
  /**
     * @minLength 1
     * @maxLength 30
     */
  displayName?: string;
  quietMode?: boolean;
}